Update Settings
Added the capability to view the properties file (app.properties) as a
menu option. Furthermore, we can make changes. This allows us to set a
different instance ID so we are no longer constrained to a single
instance.

Also, I did some refactoring and gave the GUI elements better names.